refactor(explorer): migrate MATRIX_ELEMENTS to PanelBuilder registry
Phase 2/8 (3rd of 9 PlotTypes total).

* New explorer_internals/matrix_elements.py hosts
  MatrixElementsPanelBuilder + the absorbed
  display_matrixelements render function.
* Registry adds PlotType.MATRIX_ELEMENTS; the matching
  dispatch branches are removed from Explorer.build_panel and
  ExplorerSettings.build_settings_ui.
* Existing isinstance(subsys, QubitBaseClass) dispatch guard
  becomes an assert inside the builder (same treatment as
  WAVEFUNCTIONS).
